small fixes

+++ apps/lsp_with_downcast.cpp
@@ -67,7 +67,7 @@ public:
 
   [[nodiscard]] iface::color     color() const override { return col_; }
   [[nodiscard]] std::string_view color_str() const override {
-    return iface::color_str[static_cast<int>(col_)];
+    return iface::color_str[static_cast<std::size_t>(col_)];
   }
 
 private:
